requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
accelerate>=0.27.0
kagglehub[pandas-datasets]>=0.3.0
mediapipe>=0.10.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
einops>=0.7.0
